rename to syntax_highlight.js
