+++ lib/extend/index.js
@@ -5,7 +5,7 @@ const Deployer = require('./deployer');
 const Filter = require('./filter');
 const Generator = require('./generator');
 const Helper = require('./helper');
-const Highlight = require('./highlight');
+const Highlight = require('./syntax_highlight');
 const Injector = require('./injector');
 const Migrator = require('./migrator');
 const Processor = require('./processor');
