+++ lib/extend/syntax_highlight.js
